rename: product type 모듈 src/types로 수정
