+++ src/pages/HomePage/types/index.ts
@@ -1,3 +1,5 @@
+import { ProductData } from '@/types/productType';
+
 export type RankingFilter = {
   targetType: 'ALL' | 'FEMALE' | 'MALE' | 'TEEN';
   rankType: 'MANY_WISH' | 'MANY_RECEIVE' | 'MANY_WISH_RECEIVE';
@@ -6,32 +8,6 @@ export type RankingFilter = {
 export type GetProductRankingRequest = {
   targetType?: RankingFilter['targetType'];
   rankType?: RankingFilter['rankType'];
-};
-
-interface Wish {
-  wishCount: number;
-  isWished: boolean;
-}
-
-interface Price {
-  basicPrice: number;
-  discountRate: number;
-  sellingPrice: number;
-}
-
-interface BrandInfo {
-  id: number;
-  name: string;
-  imageURL: string;
-}
-
-export type ProductData = {
-  id: number;
-  name: string;
-  imageURL: string;
-  wish: Wish;
-  price: Price;
-  brandInfo: BrandInfo;
 };
 
 export type GetRankingResponse = {
